--- src/X10A/arrayFunctions.hpp
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <Arduino.h>
-
-bool contains(uint8_t *array, size_t size, uint8_t value);
