+++ src/X10A/x10a.cpp
@@ -15,6 +15,16 @@ RegistryBuffer *registryBuffers; // holds the registries to query and the last r
 ulong lastTimeRunned = 0;
 HandleState handleState = HandleState::Stopped;
 
+bool contains(uint8_t *array, size_t size, uint8_t value)
+{
+  for (size_t i = 0; i < size; i++) {
+    if (array[i] == value) {
+      return true;
+    }
+  }
+  return false;
+}
+
 byte getCRC(byte *src, int len)
 {
   byte b = 0;
